Adding a firstName field to User
Start in our User class. This is a normal entity... so we can add whatever fields we want, like private . I'll go to Code->Generate, or Command+N on a Mac, then select 'ORM Annoations' to annotate firstName. Then I'll go back to Code->Generate to add the getter and setter methods.

Notice, right now, firstName is not nullable... meaning it's required in the database... and that's fine! If firstName is optional in your app, you can of course add nullable=true. I'm mentioning this for one reason: if you add any required fields to your User, the fos:user:create command will no longer work... because it will create a new User but leave those fields blank. I never use that command in production anyways, but, you've been warned!