Auto-Setting the Username Field
Start in the User class. I'll go to the Code->Generate menu - or Command+N on a Mac - go to 'Override Methods' and choose setEmail(). Before the parent call, add {dollar_sign}this->setUsername({dollar_sign}email)

This is big! Thanks to this, we no longer need to worry about ever setting the username field. In the database, username will always equal the email... which is definitely redundant and unnecessary, but in practice, it works fine.