Removing the Username Field from the Form
The only other thing we need to do is remove the username field from the registration form. How? Easy: in RegistrationFormType, add ->remove('username')

Then, in the template, remove its form_row()

And yea, that's it! Refresh! No more username! Register as aquanaut4@gmail.com and submit. Check it out in the database:

php bin/console doctrine:query:sql 'SELECT * FROM user'
There it is! The username is now equal to the email.

Oh, and if you're using the profile edit page from the bundle, you'll also want to remove the username field from the ProfileFormType form. It's done in exactly the same way.