We can override templates. We can override translations. We can override forms. But there's more than that. For example, after we finish registration, we're redirected to this registration confirmation page. You know what? I'd rather do something else: I'd rather redirect to the homepage and skip this page entirely. How can we do that?
Well, let's do a little bit of digging. If you hover over the route name in the web debug toolbar, you can see that this is page rendered by RegistrationController. Back in my editor I'll press Shift+Shift and look for RegistrationController in the bundle. Specifically, registerAction() is responsible for both rendering the registration page and processing the form submit.

And check this out: after the form is valid, it redirects to the confirmation page.

Events to the Rescue!

So at first, it seems like we need to override the controller itself. But not so fast! The controller - in fact every controller in FOSUserBundle is littered with events: REGISTRATION_INITIALIZE, REGISTRATION_SUCCESS, REGISTRATION_COMPLETED and REGISTRATION_FAILURE. Each of these represents a hook point where we can add custom logic.

In this case, if you look closely, you can see that after it dispatches an event called REGISTRATION_SUCCESS, below, it checks to see if the  has a response set on it. If it does not, it redirects to the confirmation page. But if it does, it uses that response.

That's the key! If we can add a listener to REGISTRATION_SUCCESS, we can create our own RedirectResponse and set that on the event so that this controller uses it. Let's go!