Creating the Event Subscriber Inside of AppBundle, create a new directory called EventListener. And in there, a new PHP class: how about RedirectAfterRegistrationSubscriber. Make this implement EventSubscriberInterface: the interface that all event subscribers must have. I'll use our favorite Code->Generate menu, or Command+N on a Mac, go to 'Implement Methods' and select getSubscribedEvents.
We want to attach a listener to FOSUserEvents::REGISTRATION_SUCCESS, which, by the way, is just a constant that equals some string event name.

In getSubscribedEvents(), add FOSUserEvents::REGISTRATION_SUCCESS assigned to onRegistrationSuccess. This means that when the REGISTRATION_SUCCESS event is fired, the onRegistrationSuccess method should be called. Create that above: public function onRegistrationSuccess().

Oh, and notice that when this event is dispatched, the bundle passes a FormEvent object. That will be the first argument to our listener method: FormEvent . That's what we need to set the response onto.

Investigating all the Events

Before we go any further, I'll hold command and click into the FOSUserEvents class... cause it's awesome! It holds a list of every event dispatched by FOSUserBundle, what its purpose is, and what event object you will receive. This is gold.