Creating the RedirectResponse Back in onRegistrationSuccess, we need to create a RedirectResponse and set it on the event. But to redirect to the homepage, we'll need the router. At the top of the class, create public function __construct() with a RouterInterface  argument. Next, I'll hit Option+Enter, select 'Initialize Fields' and choose router.
That was just a shortcut to create the private  property and set it in the constructor: nothing fancy.

Now, in onRegistrationSuccess() we can say {dollar_sign}url = {dollar_sign}this->router->generate('homepage'), and {dollar_sign}response = new RedirectResponse({dollar_sign}url). You may or may not be familiar with RedirectResponse. In a controller, to redirect, you use {dollar_sign}this->redirectToRoute(). In reality, that's just a shortcut for these two lines!

Finally, add {dollar_sign}event->setResponse({dollar_sign}response).