Ok, this class is perfect! To tell Symfony about the event subscriber, head to app/config/services.yml. At the bottom, add app.redirect_after_registration_subscriber, set the class, and add autowire: true. By doing that, thanks to the RouterInterface type-hint, Symfony will automatically know to pass us the router.
Finally, add a tag on the bottom: name: kernel.event_subscriber. And we are done!

Try it out! Go back to /register and signup as aquanaut5@gmail.com. Fill out the rest of the fields and submit!

Boom! Back to our homepage! You can customize just about anything with events. So don't override the controller. Instead, hook into an event!